Increase fluid and bin storage related configs to longs

+++ src/main/java/mekanism/common/attachments/containers/fluid/FluidTanksBuilder.java
@@ -3,6 +3,7 @@ package mekanism.common.attachments.containers.fluid;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.IntSupplier;
+import java.util.function.LongSupplier;
 import java.util.function.Predicate;
 import mekanism.api.container.LargeResourceStack;
 import mekanism.api.functions.ConstantPredicates;
@@ -32,31 +33,31 @@ public class FluidTanksBuilder {
         return new BaseFluidTankCreator(tankCreators);
     }
 
-    public <VANILLA_INPUT extends RecipeInput, RECIPE extends MekanismRecipe<VANILLA_INPUT>, INPUT_CACHE extends IInputRecipeCache> FluidTanksBuilder addBasic(int capacity,
+    public <VANILLA_INPUT extends RecipeInput, RECIPE extends MekanismRecipe<VANILLA_INPUT>, INPUT_CACHE extends IInputRecipeCache> FluidTanksBuilder addBasic(long capacity,
           IMekanismRecipeTypeProvider<VANILLA_INPUT, RECIPE, INPUT_CACHE> recipeType, ContainsRecipe<INPUT_CACHE, FluidResource> containsRecipe) {
         return addBasic(capacity, fluid -> containsRecipe.check(recipeType.getInputCache(), null, fluid));
     }
 
-    public FluidTanksBuilder addBasic(int capacity, Predicate<FluidResource> isValid) {
+    public FluidTanksBuilder addBasic(long capacity, Predicate<FluidResource> isValid) {
         return addBasic(() -> capacity, isValid);
     }
 
-    public FluidTanksBuilder addBasic(IntSupplier capacity, Predicate<FluidResource> isValid) {
+    public FluidTanksBuilder addBasic(LongSupplier capacity, Predicate<FluidResource> isValid) {
         return addTank((type, attachedTo, containerIndex) -> new ComponentBackedFluidTank(attachedTo,
               containerIndex, ConstantPredicates.manualOnly(), ConstantPredicates.alwaysTrueBi(), isValid, MekanismConfig.general.fluidItemFillRate, capacity));
     }
 
-    public FluidTanksBuilder addBasic(int capacity) {
+    public FluidTanksBuilder addBasic(long capacity) {
         return addBasic(() -> capacity);
     }
 
-    public FluidTanksBuilder addBasic(IntSupplier capacity) {
+    public FluidTanksBuilder addBasic(LongSupplier capacity) {
         return addTank((type, attachedTo, containerIndex) -> new ComponentBackedFluidTank(attachedTo,
               containerIndex, ConstantPredicates.manualOnly(), ConstantPredicates.alwaysTrueBi(), ConstantPredicates.alwaysTrue(),
               MekanismConfig.general.fluidItemFillRate, capacity));
     }
 
-    public FluidTanksBuilder addBasicExtractable(IntSupplier rate, IntSupplier capacity, Predicate<FluidResource> isValid) {
+    public FluidTanksBuilder addBasicExtractable(IntSupplier rate, LongSupplier capacity, Predicate<FluidResource> isValid) {
         return addTank((type, attachedTo, containerIndex) -> new ComponentBackedFluidTank(attachedTo,
               containerIndex, ConstantPredicates.alwaysTrueBi(), ConstantPredicates.alwaysTrueBi(), isValid, rate, capacity));
     }
